End of section 18  (App) confirmation service

+++ ClientApp/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Message } from "../_models/message";
 import { Pagination } from "../_models/pagination";
 import { MessageService } from "../_services/message.service";
+import { ConfirmService } from "../_services/confirm.service";
 
 @Component({
   selector: 'app-messages',
@@ -16,7 +17,8 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) {
+  constructor(private messageService: MessageService,
+              private confirmService: ConfirmService) {
   }
 
   ngOnInit(): void {
@@ -33,9 +35,14 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id:number){
-    this.messageService.deleteMessage(id).subscribe(()=>{
-      this.messages.splice(this.messages.findIndex(m => m.id == id),1);
-    })
+    this.confirmService.confirm('Confirm delete message', 'this cannot be undone').subscribe((result)=>{
+      if(result){
+        this.messageService.deleteMessage(id).subscribe(()=>{
+          this.messages.splice(this.messages.findIndex(m => m.id == id),1);
+        });
+      }
+    });
+
   }
 
   pageChanged(event: any) {
